Transfer login logic from AdminMain to LoginOrSignUp Try default logging in whenever user visits /admin Show AdminMain only when user is logged in & isApproved
