+++ src/components/Admin/LoginForm.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const LoginForm = () => {
+  return <div>LoginForm</div>;
+};
+
+export default LoginForm;
